Update ft_atol with long as sign to be sure !

+++ convert/ft_atol.c
@@ -6,7 +6,7 @@
 /*   By: albillie <albillie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/11 02:20:35 by albillie          #+#    #+#             */
-/*   Updated: 2024/12/11 02:28:20 by albillie         ###   ########.fr       */
+/*   Updated: 2024/12/11 04:58:47 by albillie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 long	ft_atol(const char *str)
 {
 	long	result;
-	int		sign;
+	long		sign;
 	int		i;
 
 	result = 0;
